refactor: sheet 정보 컴포넌트 분리
