+++ front/src/typings/domain.ts
@@ -43,4 +43,14 @@ type CrewList = { status: string; crews: Crew[] };
 
 type CrewListMap = Record<string, Crew[]>;
 
-export { Crew, Coach, MonthYear, Schedule, DaySchedule, ReservationInfo, CrewList, CrewListMap };
+export {
+  Crew,
+  Coach,
+  MonthYear,
+  Schedule,
+  DaySchedule,
+  ReservationInfo,
+  CrewList,
+  CrewListMap,
+  Sheets,
+};
